requirements.txt
zeep>=4.1.0
requests~=2.32.0
python-dotenv~=1.0.0
beautifulsoup4~=4.12.3
dateparser~=1.2.0